requirements.txt
python-dotenv==1.0.0
langchain==0.3.17
langchain_chroma==0.2.1
langchain_community==0.3.16
langchain_openai==0.3.6
langgraph==0.2.69
openai==1.61.0
langchain_chroma
tiktoken==0.8.0
duckduckgo-search==7.3.0
numexpr==2.10.2
wikipedia==1.4.0
streamlit==1.43.0